VERY VERY basic placeholder species descriptions
I want to elaborate more on these but that requires more lore to be refined and unfortunately that just isn't the case. It's probably best to at least tell new players what these are supposed to be; though.

Also changes the plural version of Vulpkanin to Vulpkane; a made-up word that only the ruskies have also decided to use