More trait distribution
Slugcats and the Tajara both have received traits I feel makes relative sense for them. The tajara are really close to felinids and I CONSIDERED giving them roundstart minor night vision but i want the downside for that to be substantial. the upsides aquatics have rn are in the same boat but i rule of cool'd that; this doesn't have as much wriggle room

On the opposite end of the spectrum; skugs are some weird middleground between aquatics and tajara. I still want to rename them before we get this into players' hands.

The VULPKANE are more interesting because nothing really aligns with them functionally yet so we're gonna have to think harder for things to give them to differentiate them