Implements Effigy-Specific Default Jukebox Music
I'm actually tempted to take a note from a certain megacorp and change my own name out for an in-universe one for these - but the question of "wait; the dreamer's motif is diegetic?" unfortunately dampens that a little.

Additionally standardizes effigy's maps as having jukeboxes at the bar roundstart; for the bartender's control, of course.