The Return Of The Dorms Vendor
Readded Dorms Vendor, Renamed it from LustWish to Keepsake Enticements. Didn't change the sprite because it still actually fits; and thus OrionTheFox (who is not involved in this directly) is credited. See the original PR here: https://github.com/Skyrat-SS13/Skyrat-tg/pull/18778
That said the vending refill was resprited because the old version was... does EVERYTHING have to use that latex palatte? oml

As of right now it only contains the new collars and will be replaced by a kirbyplant if you disable ERP items via config; which is the other feature of this commit.

Co-Authored-By: OrionTheFox <76465278+OrionTheFox@users.noreply.github.com>