Adds oversized/big support in the form of background resizing (#140)
## About The Pull Request
Does as the title says, if you are over 110% height or have the
oversized quirk, the background of the character creator will get bigger
to accommodate that. This is a partial port from
https://github.com/Bubberstation/Bubberstation/pull/3015. This does NOT
port the customizable backgrounds, and instead only brings over the
default all black background.
## Why It's Good For The Game
It makes creating big or oversized quirked characters way easier.
## Changelog
:cl:
qol: The background of the character creator now scales with your
character's size.
/:cl: