Tweak how loading screen fluff text is sent (#230)
## About The Pull Request

After #229 I noticed that sometimes the fluff text would stick around
after loading was finished, possibly because the master SS was calling
`send_init_fluff_message` after `COMSIG_TICKER_ENTER_PREGAME` had fired,
so I redid how fluff text is handled to be a signal instead of a for
loop through all objects thing. This has the benefit of allowing the
text object to unsubscribe from updates once pregame starts, fixing the
previously mentioned thing, but also being cleaner in general. Also
changed where fluff text is loaded (`OnConfigLoad` instead of
`Initialize`) so it's available earlier in init and changed the
`Initialize` override on the fluff text object to be a `New` override so
it can start updating even before the Atoms SS inits

## Why It's Good For The Game

Cleaner code that should be less brittle

## Changelog

N/A